Separate the main file
